Custom display now works with cell based world
Made it so custom centered displays are supported again.  Added code to the world cordinate class to provide the overal x and y cord of the world cordinate class including the cells.  As such the world can now be as big as a signed int.  If the world needs to be bigger the code can easily be changed to a long.